+++ BuildMasterAutomation/Functions/Remove-BMServer.ps1
@@ -6,9 +6,12 @@ function Remove-BMServer
     Removes a server from BuildMaster.
 
     .DESCRIPTION
-    The `Remove-BMServer` function removes a server from BuildMaster. Pass the name of the server to remove to the `Name` parameter. If the server doesn't exist, nothing happens.
+    The `Remove-BMServer` function removes a server from BuildMaster. Pass the name of the server to remove to the
+    `Name` parameter. If the server doesn't exist, an error is written. To ignore if the server exists or not, set the
+    `ErrorAction` parameter to `Ignore`.
 
-    Pass the session to the BuildMaster instance where you want to delete the server to the `Session` parameter. Use `New-BMSession` to create a session object.
+    Pass the session to the BuildMaster instance where you want to delete the server to the `Session` parameter. Use
+    `New-BMSession` to create a session object.
 
     This function uses BuildMaster's infrastructure management API.
 
@@ -20,17 +23,19 @@ function Remove-BMServer
     .EXAMPLE
     Get-BMServer -Session $session -Name 'example.com' | Remove-BMServer -Session $session
 
-    Demonstrates that you can pipe the objects returned by `Get-BMServer` into `Remove-BMServer` to remove those servers.
+    Demonstrates that you can pipe the objects returned by `Get-BMServer` into `Remove-BMServer` to remove those
+    servers.
     #>
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSShouldProcess", "")]
     [CmdletBinding(SupportsShouldProcess)]
     param(
-        [Parameter(Mandatory)]
         # The instance of BuildMaster to connect to.
-        [object]$Session,
+        [Parameter(Mandatory)]
+        [Object] $Session,
 
-        [Parameter(Mandatory,ValueFromPipelineByPropertyName)]
         # The name of the server to remove.
-        [string]$Name
+        [Parameter(Mandatory, ValueFromPipelineByPropertyName)]
+        [String] $Name
     )
 
     process
@@ -38,7 +43,17 @@ function Remove-BMServer
         Set-StrictMode -Version 'Latest'
         Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
 
-        $encodedName = [uri]::EscapeDataString($Name)
-        Invoke-BMRestMethod -Session $Session -Name ('infrastructure/servers/delete/{0}' -f $encodedName) -Method Delete
+        $server = Get-BMServer -Session $session -Name $Name -ErrorAction Ignore
+        if (-not $server)
+        {
+            $msg = "Could not delete server ""$($Name)"" because it does not exist."
+            Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+            return
+        }
+
+        $encodedName = [Uri]::EscapeDataString($Name)
+        Invoke-BMRestMethod -Session $Session `
+                            -Name ('infrastructure/servers/delete/{0}' -f $encodedName) `
+                            -Method Delete
     }
 }